pyproject.toml
[project]
name = "yamlex"
version = "1.2.0"
description = "Dynatrace Extensions: assemble extension.yaml from individual parts"
keywords = ["dynatrace", "cli", "extensions"]
readme = "README.md"
authors = [
    { name = "Vagiz Duseev", email = "vagiz.duseev@dynatrace.com" }
]
maintainers = [
    { name = "Vagiz Duseev", email = "vagiz.duseev@dynatrace.com" }
]
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.9"
dependencies = [
    "deepdiff>=8.4",
    "ruamel-yaml>=0.18",
    "typer>=0.9",
]

[project.urls]
Documentation = "https://github.com/dynatrace-extensions/dt-extensions-yamlex"
Repository = "https://github.com/dynatrace-extensions/dt-extensions-yamlex"
Issues = "https://github.com/dynatrace-extensions/dt-extensions-yamlex/issues"

[project.scripts]
yamlex = "yamlex.cli.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "bump2version>=1.0.1",
    "ruff>=0.11.10",
]

[tool.ruff]
line-length = 80
